+++ src/main/java/com/mo/mediaodyssey/socialFeature/models/DTO/FriendRequestDTO.java
@@ -2,17 +2,17 @@ package com.mo.mediaodyssey.socialFeature.models.DTO;
 
 public class FriendRequestDTO {
 
-    private Integer requestId;
+    private Long requestId;
     private Long senderId;
     private String senderEmail;
 
-    public FriendRequestDTO(Integer requestId, Long senderId, String senderEmail) {
+    public FriendRequestDTO(Long requestId, Long senderId, String senderEmail) {
         this.requestId = requestId;
         this.senderId = senderId;
         this.senderEmail = senderEmail;
     }
 
-    public Integer getRequestId() { return requestId; }
+    public Long getRequestId() { return requestId; }
     public Long getSenderId() { return senderId; }
     public String getSenderEmail() { return senderEmail; }
 }
