+++ src/main/java/com/mo/mediaodyssey/socialFeature/models/Friendship.java
@@ -9,28 +9,28 @@ public class Friendship {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name="user_id", nullable = false)
-    private Integer userId; // sender
+    private Long userId; // sender
 
     @Column(name="friend_id", nullable = false)
-    private Integer friendId; // recipient
+    private Long friendId; // recipient
 
     @Column(nullable=false)
     private boolean accepted = false;
 
     public Friendship() {}
 
-    public Friendship(Integer userId, Integer friendId) {
+    public Friendship(Long userId, Long friendId) {
         this.userId = userId;
         this.friendId = friendId;
     }
 
 
-    public Integer getId() { return id; }
-    public Integer getUserId() { return userId; }
-    public Integer getFriendId() { return friendId; }
+    public Long getId() { return id; }
+    public Long getUserId() { return userId; }
+    public Long getFriendId() { return friendId; }
     public boolean isAccepted() { return accepted; }
     public void setAccepted(boolean accepted) { this.accepted = accepted; }
 }
